--- a/test/test_aqua.jl
+++ b/test/test_aqua.jl
@@ -5,7 +5,7 @@ using FillArrays
 
 @testset "aqua" begin
     Aqua.test_all(QuantumOpticsBase;
-                  ambiguities=(exclude=[FillArrays.reshape],),  # Due to https://github.com/JuliaArrays/FillArrays.jl/issues/105#issuecomment-1518406018
+                  ambiguities=(recursive=false),
                   piracy=(broken=true,)
                   )
     # manual piracy check to exclude identityoperator