I recently encountered some numerical instability in the transform function between position and fock bases. This should reproduce it:

```julia
using QuantumOptics, Plots

b = FockBasis(99)
bp = PositionBasis(-20, 20, 20000)
ket = fockstate(b, 75);
Tpf = transform(bp, b)
plot(range(bp.xmin, bp.xmax, bp.N), real.((Tpf*ket).data))
```
This produces the following plot:
![hermite-bad](https://github.com/qojulia/QuantumOpticsBase.jl/assets/8754988/2a2c54dc-e6d8-4bfc-bf20-6b7783de84e8)

This just supposed to be the 76th hermite polynomial rescaled by some factors however one can see the significant numerical artifacts. They're also present for any larger fock states.

This prompted me to investigate, and I think I ruled out the `horner` function in QuantumOpticsBase's `polynomials.jl` file (by using the more numerically precise polynomial evaluation function in [JuliaMath/Polynomials.jl](https://github.com/JuliaMath/Polynomials.jl)). I couldn't figure out where exactly in the `hermite.A` function this numerical instability was coming from but I did find I got similar numerical instability from implementing this using Polynomials.jl with the following

```julia
import QuantumOpticsBase: transform
import Polynomials: Polynomial

function hermite_fn(N::T) where T
    A = Vector{Polynomial}(undef, N+1)
    x = Polynomial([0., 1.])
    A[1] = Polynomial([1.])
    A[2] = sqrt(2)*x
    for n=2:N
        A[n+1] = sqrt(2/n)*x*A[n] - sqrt((n-1)/n)*A[n-1]
    end
    A
end

function transform(::Type{S}, b1::PositionBasis, b2::FockBasis) where S
    T = Matrix{S}(undef, length(b1), length(b2))
    xvec = samplepoints(b1)
    A = hermite_fn(b2.N)
    delta_x = spacing(b1)
    c = 1.0/sqrt(sqrt(pi))*sqrt(delta_x)
    for i in 1:length(b1)
        u = xvec[i]
        C = c*exp(-u^2/2)
        for n=b2.offset:b2.N
            idx = n-b2.offset+1
            T[i,idx] = C*A[n+1](u)
        end
    end
    DenseOperator(b1, b2, T)
end
```

I also tried using `Hermite` from [jverzani/SpecialPolynomials.jl](https://github.com/jverzani/SpecialPolynomials.jl) however I still saw this numerical instability.

It seems that [ClassicalOrthogonalPolynomials.jl](https://github.com/JuliaApproximation/ClassicalOrthogonalPolynomials.jl) is somehow able to avoid this numerical instability, although I haven't figured out exactly how since the code in this package is pretty advanced. Anyways, this PR does fix this numerical instability, although it does add in quite a few dependencies from ClassicalOrthogonalPolynomials.jl. Also the change in `ptranspose` is a workaround for a bug I found in QuasiArrays.jl: (https://github.com/JuliaApproximation/QuasiArrays.jl/issues/97). Hence why I'm opening this as a draft since I'm not sure if this is the right approach to fixing this here.