@Krastanov 2023-09-12
This is some great investigative work!

The only worry I have is that the new dependency is a bit young and does not have many dependents yet: https://juliahub.com/ui/Packages/General/ClassicalOrthogonalPolynomials

The new dependency is quite healthy though, belonging to the JuliaApproximations organization, with the same developers as ApproxFun, so it is probably high quality. As long as it does not have long TTFX ;)

I am a bit on the fence because of the dependency, but leaning towards accepting this. I will wait on a final verdict by @david-pl 

PS: This is always nice to see:
![image](https://github.com/qojulia/QuantumOpticsBase.jl/assets/705248/bbfeaf28-694d-46f7-bc78-9fdd0739cc7e)

It might be a new dependency, but it does remove a ton of bespoke code that we do not need to support anymore. Instead, we now use code written by domain experts.