@Krastanov 2023-09-12
The test failures come from Aqua detecting a very large number of method ambiguities in the newly imported dependencies. That is not necessarily a show stopper for us, we already exclude some packages from ambiguity detection [see here](https://github.com/qojulia/QuantumOpticsBase.jl/blob/master/test/test_aqua.jl#L8), but it is something that should at least be on the radar of the developers of the dependency.

@akirakyle , could you
- blacklist the packages so that aqua does not complain about them anymore (see the link above for an example -- you should be able to blacklist whole packages, no need to blacklist separate functions (that would be very teddious and fragile))
- reach out to the issue trackers of the various blacklisted packages and check whether they are aware or plan to do anything about it

Method ambiguities can go from harmless to causing significant TTFX on a small change, so we should be careful.

Other tests seem to run fine, so nothing is broken here.