@akirakyle 2023-09-12
Thanks! That's good to hear, I'm still never quite sure how much to trust various julia packages given both that I'm new to the ecosystem and that the ecosystem is still relatively new, but I do agree that in long term it will be good to move these bespoke functions to domain specific packages like this one.

Also I forgot to add that I did some timings. For

```julia
@time Tpf = transform(bp, b)
```

Before this PR I get for the first run
```
 0.213219 seconds (128.41 k allocations: 39.285 MiB, 7.49% gc time, 42.21% compilation time)
```
Then subsequent runs I get
```
0.151279 seconds (110 allocations: 30.717 MiB, 8.53% gc time)
```

After this PR I get first run
```
1.012682 seconds (12.94 M allocations: 450.505 MiB, 7.84% gc time, 78.97% compilation time)
```
Then subsequent runs I get
```
0.232888 seconds (9.80 M allocations: 240.023 MiB, 18.45% gc time)
```

Which I honestly find quite impressive as it seems ClassicalOrthogonalPolynomials.jl has a lot of useful abstractions and fixes numerical instability with seemingly very little cost!