@akirakyle 2023-09-14
>    * blacklist the packages so that aqua does not complain about them anymore (see the link above for an example -- you should be able to blacklist whole packages, no need to blacklist separate functions (that would be very teddious and fragile))

Unfortunately it doesn't appear there's a way to remove entire packages from the ambiguity test (https://juliatesting.github.io/Aqua.jl/dev/#Aqua.test_ambiguities-Tuple{Any}). Instead I just disabled recursive checking of ambiguities. Another possibility would be to keep recursive ambiguity testing but remove checks for ambiguities with Base via
```julia
Aqua.test_ambiguities([QuantumOpticsBase, Core]; exclude=[Infinities.ComplexInfinity])
```
There's a relevant issue in the Aqua.jl repo discussing this here: https://github.com/JuliaTesting/Aqua.jl/issues/77

>    * reach out to the issue trackers of the various blacklisted packages and check whether they are aware or plan to do anything about it

There's 9 direct dependencies of ClassicalOrthogonalPolynomials and two transitive dependencies (Infinities and SemiseparableMatrices) with ambiguity issues. Given they're all under various Julia* github organizations and it appears they share developers in common I went ahead and just opened an issue ClassicalOrthogonalPolynomials explaining the situation (https://github.com/JuliaApproximation/ClassicalOrthogonalPolynomials.jl/issues/154).