@dlfivefifty 2023-09-14
Note `OrthonormalWeighted(Normalized(Hermite()))` still has issues with underflow:
```julia
julia> norm(H[40,1:100])
0.0
```
There's a much better way to do this: take the n-th root  `w_n = exp(-x^2/(2n))` and incorporate it as you build the recurrence. This is something I haven't got around to but there is some version of this method here:

https://github.com/JuliaApproximation/FastGaussQuadrature.jl/blob/a5932947c43186d1e13730015da705b5052012f5/src/gausshermite.jl#L130


Given your concerns with ambiguities and since you only need evaluation I'm wondering if a separate package (OrthogonalPolynomialRecurrences.jl) would be worthwhile? The design of ClassicalOrthogonalPolynomials.jl depends on InfiniteArrays.jl to avoid rewriting code for each family but this can be avoided by just rewriting the code for each family.