@dlfivefifty 2023-09-14
Or just use FastTransforms.jl? This is the relationship:
```julia
julia> OrthonormalWeighted(Hermite())[x, 1:n]  ≈ FastGaussQuadrature.hermpoly_rec(0:n-1, sqrt(2)* x) * π^(-1/4)
true
```
(So you would also need to add the normalisation constant.) But then this avoids underflow:
```julia
julia> x = 40; n = 100; FastGaussQuadrature.hermpoly_rec(0:n-1, sqrt(2)* x) * π^(-1/4)
100-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 ⋮
 9.961817500563452e-262
 5.804365934206908e-261
 3.361991950596754e-260
 1.9359355316676075e-259
 1.1083157837936175e-258
 6.308702574489045e-258
 3.570634917149043e-257
 2.0095745538076928e-256
 1.1247085178318906e-255
 6.2600303429468115e-255
 3.4652646464337584e-254
 1.9078417421194034e-253
```