@david-pl 2023-09-14
@Krastanov thanks for the ping. The dependency looks solid to me, so feel free to merge as soon as you're happy with the code.

@akirakyle thanks for the deep dive you did here ;)