@akirakyle 2023-09-14
> Or just use FastGaussQuadrature.jl?

Thanks @dlfivefifty for pointing out `hermpoly_rec` in FastGaussQuadrature.jl! I somehow missed that when I was looking through it and thought from the documentation that FastGaussQuadrature.jl only implemented functions that efficiently found the roots of the hermite polynomials without actually needing to evaluate them directly.

> (So you would also need to add the normalisation constant.) But then this avoids underflow:

Ah I also missed this underflow in my testing. Thanks for explicitly giving the relationship between `Hermite` and `hermpoly_rec`, I've created a new PR at #142 that should be preferred to this one.